# managed/gateway/openapi/managed.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Steel Managed Operations API",
    "version": "2026-07-01",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/v1/managed/admissions": {
      "post": {
        "operationId": "createManagedAdmission",
        "tags": [
          "Managed operations"
        ],
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagedAdmissionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful managed operation",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admission"
                }
              }
            }
          },
          "400": {
            "description": "Managed error (400)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "INVALID_ARGUMENT"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Managed error (401)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "ACCESS_AUTH_REQUIRED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Managed error (409)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "CREATE_TOKEN_CONFLICT"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Managed error (429)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "ADMISSION_TIMEOUT",
                                "QUEUE_FULL",
                                "RATE_LIMITED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "Managed error (502)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "UPSTREAM_BAD_RESPONSE"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Managed error (503)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "NO_REACHABLE_WORKER",
                                "MANAGER_DRAINING",
                                "MANAGED_LIMITER_CAPACITY",
                                "MANAGED_TICKET_CAPACITY",
                                "MANAGED_IDEMPOTENCY_CAPACITY"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "504": {
            "description": "Managed error (504)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "UPSTREAM_TIMEOUT"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/managed/admissions/{id}/cancel": {
      "post": {
        "operationId": "cancelManagedAdmission",
        "tags": [
          "Managed operations"
        ],
        "security": [
          {
            "cloudflareAccess": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyManagedMutationBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful managed operation",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admission"
                }
              }
            }
          },
          "400": {
            "description": "Managed error (400)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "INVALID_ARGUMENT"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Managed error (401)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "ACCESS_AUTH_REQUIRED"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Managed error (404)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "ADMISSION_NOT_FOUND"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Managed error (409)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "SESSION_STATE_CONFLICT"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "502": {
            "description": "Managed error (502)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "UPSTREAM_BAD_RESPONSE"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "504": {
            "description": "Managed error (504)",
            "headers": {
              "X-Managed-Api-Version": {
                "required": true,
                "schema": {
                  "type": "string",
                  "enum": [
                    "2026-07-01"
                  ]
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ManagedErrorEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "object",
                          "properties": {
                            "code": {
                              "type": "string",
                              "enum": [
                                "UPSTREAM_TIMEOUT"
                              ]
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cloudflareAccess": {
        "type": "apiKey",
        "in": "header",
        "name": "Cf-Access-Jwt-Assertion"
      }
    },
    "schemas": {
      "Admission": {
        "type": "object",
        "properties": {
          "admissionId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "state": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RESERVED",
              "STARTING",
              "ADMITTED",
              "CANCELLED",
              "EXPIRED",
              "FAILED"
            ]
          },
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "sessionId": {
            "type": "string",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "failureCode": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{0,63}$"
          }
        },
        "required": [
          "admissionId",
          "state",
          "createdAt",
          "expiresAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "ManagedAdmissionCreateRequest": {
        "type": "object",
        "properties": {
          "idempotencyKey": {
            "type": "string",
            "pattern": "^[A-Za-z0-9._:-]{8,128}$"
          },
          "operation": {
            "type": "string",
            "enum": [
              "SESSION_CREATE"
            ]
          }
        },
        "required": [
          "idempotencyKey",
          "operation"
        ],
        "additionalProperties": false
      },
...
      "EmptyManagedMutationBody": {
        "type": "object",
        "properties": {},
        "additionalProperties": false
      },
...
      "JsonValue": {
        "oneOf": [
          {
            "enum": [
              null
            ]
          },
          {
            "type": "boolean"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            }
          },
          {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            }
          }
        ]
      },
      "ManagedErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "apiVersion",
          "error"
        ],
        "properties": {
          "apiVersion": {
            "type": "string",
            "enum": [
              "2026-07-01"
            ]
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "retryable",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INVALID_ARGUMENT",
                  "INVALID_CURSOR",
                  "ACCESS_AUTH_REQUIRED",
                  "ACCESS_FORBIDDEN",
                  "SESSION_NOT_FOUND",
                  "ADMISSION_NOT_FOUND",
                  "RESULT_NOT_FOUND",
                  "TOOL_NOT_FOUND",
                  "ROUTE_NOT_FOUND",
                  "METHOD_NOT_ALLOWED",
                  "NOT_ACCEPTABLE",
                  "MANAGED_SESSION_REQUIRED",
                  "SESSION_STATE_CONFLICT",
                  "CREATE_TOKEN_CONFLICT",
                  "EVENT_CURSOR_RESTARTED",
                  "LIST_CURSOR_RESTARTED",
                  "INSTANCE_LOST",
                  "EVENT_CURSOR_EXPIRED",
                  "LIST_CURSOR_EXPIRED",
                  "BODY_TOO_LARGE",
                  "RESULT_TOO_LARGE",
                  "UNSUPPORTED_MEDIA_TYPE",
                  "MANAGED_FEATURE_UNSUPPORTED",
                  "TOOL_INPUT_INVALID",
                  "ADMISSION_TIMEOUT",
                  "QUEUE_FULL",
                  "RATE_LIMITED",
                  "UPSTREAM_BAD_RESPONSE",
                  "NO_REACHABLE_WORKER",
                  "MANAGER_DRAINING",
                  "MANAGED_LIMITER_CAPACITY",
                  "MANAGED_TICKET_CAPACITY",
                  "MANAGED_IDEMPOTENCY_CAPACITY",
                  "MANAGED_SNAPSHOT_CAPACITY",
                  "MANAGED_INGRESS_CAPACITY",
                  "MANAGED_AI_RESULT_CAPACITY",
                  "MANAGED_ACTION_CAPACITY",
                  "MANAGED_WS_CAPACITY",
                  "WORKER_JOURNAL_CAPACITY",
                  "UPSTREAM_TIMEOUT",
                  "TOOL_TIMEOUT"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 1024
              },
              "retryable": {
                "type": "boolean"
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              },
              "details": {
                "$ref": "#/components/schemas/JsonValue"
              }
            }
          }
        }
      }
    }
  }
}